+++ test/src/custom_keyboard_test.dart
@@ -1,5 +1,4 @@
 import 'package:custom_pin_screen/pin_screen.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';