+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+## 0.1
+
+* Added pin authentication widget.
+* Added tests for pin authentication.
+
 ## 0.0.1
 
 * Added custom keyboard.
-* Added tests for custom keyboard
+* Added tests for custom keyboard.